Slight change to background and deleted Unnecessary component of CustBG
